Refactor initialization
Refactor the skill initialization to improve request response times.
