+++ src/LondonTravel.Skill/AlexaFunction.cs
@@ -6,7 +6,13 @@ using System.Threading.Tasks;
 using Alexa.NET.Request;
 using Alexa.NET.Response;
 using Amazon.Lambda.Core;
+using MartinCostello.LondonTravel.Skill.Extensions;
+using MartinCostello.LondonTravel.Skill.Intents;
+using Microsoft.ApplicationInsights;
+using Microsoft.ApplicationInsights.Extensibility;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Logging;
 
 namespace MartinCostello.LondonTravel.Skill
 {
@@ -15,6 +21,18 @@ namespace MartinCostello.LondonTravel.Skill
     /// </summary>
     public class AlexaFunction
     {
+        /// <summary>
+        /// The <see cref="IServiceProvider"/> to use.
+        /// </summary>
+        private IServiceProvider _serviceProvider;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="AlexaFunction"/> class.
+        /// </summary>
+        public AlexaFunction()
+        {
+        }
+
         /// <summary>
         /// Handles a request to the skill as an asynchronous operation.
         /// </summary>
@@ -27,11 +45,25 @@ namespace MartinCostello.LondonTravel.Skill
         {
             context.Logger.LogLine($"Invoking skill request of type {request.Request.GetType().Name}.");
 
-            IServiceProvider serviceProvider = CreateServiceProvider();
-
-            var handler = serviceProvider.GetRequiredService<FunctionHandler>();
+            var handler = _serviceProvider.GetRequiredService<FunctionHandler>();
 
             return await handler.HandleAsync(request);
+        }
+
+        /// <summary>
+        /// Initializes the skill as an asynchronous operation.
+        /// </summary>
+        /// <returns>
+        /// A <see cref="Task{TResult}"/> representing the asynchronous operation to initialize the skill.
+        /// </returns>
+        public Task<bool> InitializeAsync()
+        {
+            if (_serviceProvider == null)
+            {
+                _serviceProvider = CreateServiceProvider();
+            }
+
+            return Task.FromResult(true);
         }
 
         /// <summary>
@@ -40,7 +72,44 @@ namespace MartinCostello.LondonTravel.Skill
         /// <param name="services">The service collection to configure.</param>
         protected virtual void ConfigureServices(IServiceCollection services)
         {
-            // No-op
+            services.AddLogging((builder) => builder.AddLambdaLogger());
+
+            services.AddHttpClients();
+            services.AddPolly();
+
+            services.TryAddSingleton((_) => SkillConfiguration.CreateDefaultConfiguration());
+
+            services.AddSingleton<AlexaSkill>();
+            services.AddSingleton<FunctionHandler>();
+            services.AddSingleton<IntentFactory>();
+            services.AddSingleton((_) => TelemetryConfiguration.CreateDefault());
+            services.AddSingleton(CreateTelemetryClient);
+
+            services.AddSingleton<EmptyIntent>();
+            services.AddSingleton<HelpIntent>();
+            services.AddSingleton<UnknownIntent>();
+
+            services.AddTransient<CommuteIntent>();
+            services.AddTransient<DisruptionIntent>();
+            services.AddTransient<StatusIntent>();
+        }
+
+        /// <summary>
+        /// Creates an <see cref="TelemetryClient"/>.
+        /// </summary>
+        /// <param name="serviceProvider">The <see cref="IServiceProvider"/> to use.</param>
+        /// <returns>
+        /// The created instance of <see cref="TelemetryClient"/>.
+        /// </returns>
+        private static TelemetryClient CreateTelemetryClient(IServiceProvider serviceProvider)
+        {
+            var config = serviceProvider.GetRequiredService<SkillConfiguration>();
+            var configuration = serviceProvider.GetRequiredService<TelemetryConfiguration>();
+
+            return new TelemetryClient(configuration)
+            {
+                InstrumentationKey = config.ApplicationInsightsKey,
+            };
         }
 
         /// <summary>
@@ -51,7 +120,11 @@ namespace MartinCostello.LondonTravel.Skill
         /// </returns>
         private IServiceProvider CreateServiceProvider()
         {
-            return ServiceResolver.GetServiceCollection(ConfigureServices).BuildServiceProvider();
+            var services = new ServiceCollection();
+
+            ConfigureServices(services);
+
+            return services.BuildServiceProvider();
         }
     }
 }
