+++ test/LondonTravel.Skill.Tests/HelpTests.cs
@@ -22,7 +22,7 @@ namespace MartinCostello.LondonTravel.Skill
         public async Task Can_Invoke_Function()
         {
             // Arrange
-            AlexaFunction function = CreateFunction();
+            AlexaFunction function = await CreateFunctionAsync();
 
             SkillRequest request = CreateIntentRequest("AMAZON.HelpIntent");
             ILambdaContext context = CreateContext();
