+++ test/LondonTravel.Skill.Tests/LaunchTests.cs
@@ -23,7 +23,7 @@ namespace MartinCostello.LondonTravel.Skill
         public async Task Can_Invoke_Function()
         {
             // Arrange
-            AlexaFunction function = CreateFunction();
+            AlexaFunction function = await CreateFunctionAsync();
 
             SkillRequest request = CreateRequest<LaunchRequest>();
             ILambdaContext context = CreateContext();
